pyre-jit-trace: extract residual-call dispatch to residual_call.rs
Relocate the 17 residual-call helpers — the per-shape dispatchers
(dispatch_residual_call_{iRd,iIRd,iIRFd}_kind), the executor fast paths
(try_fold_pure_call_via_executor, try_execute_residual_call_via_executor),
opcode selection, arg binding, pre-call vable/vref sync, result writeback,
and the body classification helpers — from jitcode_dispatch/mod.rs into
residual_call.rs. Pure relocation; the residual_call_* opname arms stay in
handle and call into the moved dispatchers.

t3_audit_opname_gap_inventory now scans residual_call.rs alongside mod.rs
and arith.rs so the moved residual-call match arms stay in its
source-of-truth handler set.
